chore: gnb, bnb 메뉴 아이콘 추가
