--- src/constants/constants.ts
@@ -1,45 +1,0 @@
-export const GlobalMenus = {
-  "/": {
-    title: "",
-    firstIcon: "search",
-    secondtIcon: "heart",
-  },
-  "/challenge": {
-    title: "챌린지",
-    firstIcon: "search",
-    secondtIcon: "heart",
-  },
-  "/certification": {
-    title: "챌린지 인증",
-    firstIcon: "search",
-    secondtIcon: "heart",
-  },
-  "/mypage": {
-    title: "마이페이지",
-    firstIcon: "bell",
-    secondtIcon: "setting",
-  },
-};
-
-export const BottomMenus = [
-  {
-    url: "/",
-    title: "홈",
-    icon: "",
-  },
-  {
-    url: "/challenge",
-    title: "챌린지",
-    icon: "",
-  },
-  {
-    url: "/certification",
-    title: "인증",
-    icon: "",
-  },
-  {
-    url: "/mypage",
-    title: "마이페이지",
-    icon: "",
-  },
-];
